Add Maven Wrapper configuration for version 3.9.9

+++ gallery-ai/src/main/java/com/hawkins/gallery/service/KnownFaceService.java
@@ -100,7 +100,7 @@ public class KnownFaceService {
                 .map(row -> new KnownPersonSummary(
                         (String) row[0],
                         (String) row[1],
-                        ((Long) row[2]).intValue()))
+                        Math.toIntExact((Long) row[2])))
                 .toList();
     }
